change project to TS
