+++ api/package.json
@@ -1,14 +1,14 @@
 {
   "name": "api",
-  "type": "commonjs",
+  "type": "module",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "babel ../common -d common",
-    "start": "npm run build && node index.js",
-    "lint": "eslint --ext .js --fix ."
+    "start": "ts-node index.ts",
+    "lint": "eslint --ext .js .ts --fix ."
   },
   "author": "",
   "license": "ISC",
@@ -16,8 +16,15 @@
     "express": "^4.18.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.23.6",
+    "@babel/core": "^7.23.7",
+    "@babel/node": "^7.22.19",
+    "@babel/preset-env": "^7.23.7",
+    "@babel/preset-typescript": "^7.23.3",
+    "@types/express": "^4.17.21",
     "babel-cli": "^6.26.0",
-    "eslint": "^8.56.0"
+    "babel-plugin-module-resolver": "^5.0.0",
+    "eslint": "^8.56.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3"
   }
 }
