--- common/index.js
@@ -1,15 +1,0 @@
-/** export sum math function */
-var sum = require('./sum.js');
-
-/** export subtract math function */
-var subtract= require('./subtract.js');
-
-/** @typedef {import('../../common') } export both math functions */
-
-module.exports= {sum,subtract};
-
-/* when type module*/
-// import {sum} from './sum.js';
-// import {subtract} from './subtract.js';
-// export {sum,subtract};
-
