+++ common/index.ts
@@ -1,0 +1,13 @@
+// /** export sum math function */
+// import { sum } from "./sum";
+// /** export subtract math function */
+// import { subtract } from "./subtract";
+
+// /** @typedef {import('../../common') } export both math functions */
+
+// export { sum, subtract };
+
+/** @typedef {import('../../common') } export both math functions */
+import { Sum } from "./sum";
+import { Subtract } from "./subtract";
+export { Sum, Subtract };
