--- common/sum.js
@@ -1,8 +1,0 @@
-/** Math function subtract contain two argument, it should be number, returns sum of two number
-* @param {number} value1 some number e.g 12
-* @param {number} value2 some number e.g 15
-* @type {number}
-* @returns sum of two number
-*/
-const sum = (value1, value2) => parseFloat(value1) + parseFloat(value2);
-module.exports= sum;
