--- web/.gitignore
@@ -1,31 +1,0 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-# Logs
-logs
-*.log
-npm-debug.log*
-
-# dependencies
-/node_modules
-/.pnp
-.pnp.js
-
-# testing
-/coverage
-
-# production
-/build
-
-# misc
-.DS_Store
-.env
-
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# Optional npm cache directory
-.npm
-
-
-# build / generate output
-dist
