+++ web/README.md
@@ -1,29 +1,46 @@
-# Getting Started with React App
+# Getting Started with Create React App
 
-This is starter project, which based on React 18 and Webpack
+This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
 
-After cloning this project. You need to run below commands.
+## Available Scripts
 
-Install the node modules - 
+In the project directory, you can run:
 
-```
-npm 1nstall
-```
+### `npm start`
 
-Bundle the project - 
-```
-npm run build
-```
+Runs the app in the development mode.\
+Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
 
-Run bundle the project - 
-```
-npm run start
-```
+The page will reload if you make edits.\
+You will also see any lint errors in the console.
 
-If you want to check the preview. Simply run the below command. Open dev server -
+### `npm test`
 
-```
-npm run dev
-```
+Launches the test runner in the interactive watch mode.\
+See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
 
-Then open the URL on your browser. (http://localhost:8080/)
+### `npm run build`
+
+Builds the app for production to the `build` folder.\
+It correctly bundles React in production mode and optimizes the build for the best performance.
+
+The build is minified and the filenames include the hashes.\
+Your app is ready to be deployed!
+
+See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
+
+### `npm run eject`
+
+**Note: this is a one-way operation. Once you `eject`, you can’t go back!**
+
+If you aren’t satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
+
+Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you’re on your own.
+
+You don’t have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn’t feel obligated to use this feature. However we understand that this tool wouldn’t be useful if you couldn’t customize it when you are ready for it.
+
+## Learn More
+
+You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
+
+To learn React, check out the [React documentation](https://reactjs.org/).
