--- web/src/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Test App</title>
-</head>
-<body>
-    <!-- app for render components -->
-    <div id="root"></div>
-</body>
-</html>
